requirements.txt
# Audio processing and handling
PyAudio>=0.2.11
webrtcvad>=2.0.10
numpy>=1.21.0
pygame>=2.0.0
edge-tts>=6.1.9

# Speech recognition
funasr>=1.0.0
modelscope>=1.10.0

# Computer vision
opencv-python>=4.5.0

# Web and network
requests>=2.25.0
beautifulsoup4>=4.9.0

# Machine learning and AI
openai>=1.3.5
dashscope>=1.14.0

# Search engines
baidusearch
googlesearch-python
# Utilities